Refactor by modularizing
